Removed some controllers and renamed files.
The "therapist" and "parent" controllers have been merged into one
"users" controller - the similarity between the two types outweigh
the the few differences in attributes. The separation was/is not
needed. Also, the "match" and "schedule" controllers can be turned
into base classes and abstracted away into plain objects.

The UsersController can pull in Match and Schedule objects and work
with those in lieu of them being controllers.