Refactored entities, controllers, views.
Added an optional "mode" parameter to the "users" route to
allow for "profile/edit" to display a form for editing the
user's profile information. It could also be used to allow
for POST requests to add availabilities to the user's Schedule.

Also introduced "AbaLookup\Entity\ScheduleDay" as a link
between "AbaLookup\Entity\ScheduleInterval" and "AbaLookup\Entity\Schedule"