Login and register view cleanup/style fixes.
I can't help but shake the feeling that the logic in these two views is similar
enough to be in its own view helper. Even still, I'm reluctant to essentially
recreate Zend\Form\View\Helper\Form because that's what it would probably look like.

I welcome feedback as to what should be done here. (It's not a pressing matter.)