Putting more thought into the build process.
The only 3rd-party libs left in the repository now are the CSS/Sass
frameworks (Foundation and Normalize). Those though will not likely
change very often.

JavaScript dependencies are installed via BOWER. PHP via Composer.

The makefile in the repository is simply a way to bunch multiple
commands into a single make command - which looks cleaner in the docs.
The following targets currently exist:

               all Does nothing.
           install Installs BOWER and Grunt.
      dependencies Installs dependencies via Composer, npm, and bower.
          database Creates the database schema.
              test Runs the test script.