Refactored entities.
Refactored classes in `AbaLookup\Entity` namespace. The location specific classes
are likely no longer needed as we are going to be using a 3rd-party maps/distance
API. Added a UserType enumeration (kinda?) class that holds the types a user can be.
Refactored unit tests to reflect updates to classes.

Also,

- Removed traces of moderator flag until we figure out what that is/will be. (#52)
- Renamed sex to gender. (#63)
- Certificate of Conduct now stored as the date recieved, NULL indicating that it was not recieved. (#93)
- UserType class uses reflection (which is probably a bad idea).