Two different registration pages.
- Closes #93
- Related to #63
- Closes #62
- Closes #61

Broke the registration form into two separate pages (two separate routes),
one for parents, and one for therapists. Behind the scenes, they act the
same way, but the difference is in what fields/inputs are shown to the user.

The therapist registration form now has on it a date field to indicate the
date on which the user recieved their Certificate of Conduct. The input is
the `date` input field which will be impossible to support cross-brower w/o
using JavaScript.